% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partial_residual.R
\name{partial_residual_plot}
\alias{partial_residual_plot}
\title{Partial Residual Plot}
\usage{
partial_residual_plot(
  plot_formula,
  lm_formula = NULL,
  model = NULL,
  data,
  added_term = NULL,
  suppress_model = F,
  ...
)
}
\arguments{
\item{plot_formula}{a flexplot-style formula for plotting}

\item{lm_formula}{Optional. A lm-style formula for the model being fit}

\item{model}{Optional. A lm model}

\item{data}{The dataset}

\item{added_term}{a formula, which specifies which terms should be de-residualized. By default,
nothing is added back in (so the plot will just show the residuals of the model).}

\item{suppress_model}{A boolean. Should the visual fit of the model be suppressed? Defaults to T.}

\item{...}{Other arguments passed to flexplot}
}
\value{
a partial residual plot
}
\description{
Partial Residual Plot
}
\examples{
partial_residual_plot(weight.loss~therapy.type,
   lm_formula = weight.loss~therapy.type + motivation,
   data=exercise_data)
# prp with a model (showing regular residuals)
mod = lm(y~a + b + z + x, data=small)
partial_residual_plot(y~x, lm_formula = y~x + b, data=small)

# prp, but we add back in the x effect
partial_residual_plot(y~x, model=mod, added_term=~x, data=small)

# now show show a scatterplot of the partial residuals and overlay a loess line
partial_residual_plot(y~x, model=mod, added_term=~x, data=small, suppress_model=T, method="loess")
}
